refactor: rename tests/convert/normalize.rs
This renames the test module to match the new function name.
